Allow for copypasta to always process the rest of the message as a single argument
Remember that one time somebody was asking if there were any copypastas they should avoid and I said "//copypasta emoji I think" and it ran copypasta emoji? That was cone. This fixes that and adds support for spaces in the copypasta names. I dont think this is too unstructured for you. I even added two comment lines to document why it was there!